get-terraform-version/requirements.txt
python-hcl2==4.3.5

# The pytest framework makes it easy to write small tests,
# yet scales to support complex functional testing
pytest==8.3.3
